# Conflicts:
#	.gitignore
#	_locales/en/messages.json
#	ts/components/DebugLogView.tsx
#	ts/components/SessionWrapperModal.tsx
#	ts/components/basic/SessionHTMLRenderer.tsx
#	ts/components/basic/SessionRadio.tsx
#	ts/components/buttons/MenuButton.tsx
#	ts/components/conversation/SessionConversation.tsx
#	ts/components/conversation/SubtleNotification.tsx
#	ts/components/conversation/TimerNotification.tsx
#	ts/components/conversation/composition/CompositionBox.tsx
#	ts/components/conversation/message/message-content/MessageText.tsx
#	ts/components/conversation/message/message-item/InteractionNotification.tsx
#	ts/components/conversation/message/reactions/ReactionPopup.tsx
#	ts/components/conversation/right-panel/overlay/OverlayRightPanelSettings.tsx
#	ts/components/conversation/right-panel/overlay/disappearing-messages/DisappearingModes.tsx
#	ts/components/conversation/right-panel/overlay/disappearing-messages/OverlayDisappearingMessages.tsx
#	ts/components/dialog/BanOrUnbanUserDialog.tsx
#	ts/components/dialog/DeleteAccountModal.tsx
#	ts/components/dialog/EditProfileDialog.tsx
#	ts/components/dialog/ModeratorsAddDialog.tsx
#	ts/components/dialog/OnionStatusPathDialog.tsx
#	ts/components/dialog/ReactListModal.tsx
#	ts/components/dialog/SessionSeedModal.tsx
#	ts/components/dialog/SessionSetPasswordDialog.tsx
#	ts/components/dialog/UserDetailsDialog.tsx
#	ts/components/leftpane/LeftPaneSectionHeader.tsx
#	ts/components/leftpane/LeftPaneSettingSection.tsx
#	ts/components/leftpane/conversation-list-item/InteractionItem.tsx
#	ts/components/leftpane/overlay/OverlayClosedGroup.tsx
#	ts/components/leftpane/overlay/OverlayCommunity.tsx
#	ts/components/leftpane/overlay/OverlayMessage.tsx
#	ts/components/leftpane/overlay/SessionJoinableDefaultRooms.tsx
#	ts/components/leftpane/overlay/choose-action/ContactsListWithBreaks.tsx
#	ts/components/leftpane/overlay/choose-action/OverlayChooseAction.tsx
#	ts/components/menu/Menu.tsx
#	ts/components/registration/RegistrationStages.tsx
#	ts/components/registration/RegistrationUserDetails.tsx
#	ts/components/registration/SignInTab.tsx
#	ts/components/registration/SignUpTab.tsx
#	ts/components/settings/SessionSettings.tsx
#	ts/components/settings/SessionSettingsHeader.tsx
#	ts/components/settings/ZoomingSessionSlider.tsx
#	ts/components/settings/section/CategoryAppearance.tsx
#	ts/components/settings/section/CategoryHelp.tsx
#	ts/components/settings/section/CategoryPermissions.tsx
#	ts/components/settings/section/CategoryPrivacy.tsx
#	ts/hooks/useParamSelector.ts
#	ts/mains/main_renderer.tsx
#	ts/models/message.ts
#	ts/node/menu.ts
#	ts/node/tray_icon.ts
#	ts/session/constants.ts
#	ts/session/disappearing_messages/timerOptions.ts
#	ts/session/utils/Toast.tsx
#	ts/state/selectors/search.ts
#	ts/test/session/unit/selectors/conversations_test.ts
#	ts/types/LocalizerKeys.ts
#	ts/types/Util.ts
#	ts/window.d.ts
#	yarn.lock

+++ ts/interactions/conversations/unsendingInteractions.ts
@@ -9,12 +9,12 @@ import { SnodeAPI } from '../../session/apis/snode_api/SNodeAPI';
 import { SnodeNamespaces } from '../../session/apis/snode_api/namespaces';
 import { getConversationController } from '../../session/conversations';
 import { UnsendMessage } from '../../session/messages/outgoing/controlMessage/UnsendMessage';
-import { ed25519Str } from '../../session/onions/onionPath';
 import { PubKey } from '../../session/types';
 import { ToastUtils, UserUtils } from '../../session/utils';
 import { closeRightPanel, resetSelectedMessageIds } from '../../state/ducks/conversations';
 import { updateConfirmModal } from '../../state/ducks/modalDialog';
 import { resetRightOverlayMode } from '../../state/ducks/section';
+import { ed25519Str } from '../../session/utils/String';
 
 /**
  * Deletes messages for everyone in a 1-1 or everyone in a closed group conversation.
@@ -303,7 +303,7 @@ const doDeleteSelectedMessages = async ({
   }
 
   const isAllOurs = selectedMessages.every(message => ourDevicePubkey === message.getSource());
-  if (conversation.isPublic()) {
+  if (conversation.isPublic() && deleteForEveryone) {
     await doDeleteSelectedMessagesInSOGS(selectedMessages, conversation, isAllOurs);
     return;
   }
@@ -334,6 +334,24 @@ const doDeleteSelectedMessages = async ({
 
   // #endregion
 };
+
+/**
+ * Either delete for everyone or not, based on the props
+ */
+export async function deleteMessagesForX(
+  messageIds: Array<string>,
+  conversationId: string,
+  /** should only be enforced for messages successfully sent on communities */
+  enforceDeleteServerSide: boolean
+) {
+  if (conversationId) {
+    if (enforceDeleteServerSide) {
+      await deleteMessagesByIdForEveryone(messageIds, conversationId);
+    } else {
+      await deleteMessagesById(messageIds, conversationId);
+    }
+  }
+}
 
 export async function deleteMessagesByIdForEveryone(
   messageIds: Array<string>,
